requirements.txt
djangorestframework
markdown
django-filter
django-environ
stripe
python-environ